LILY-198 Namen/ beschrijving adres velden toevoegen
The labels for fields in the forms where overwritten to an empty string. This
was done to prevent cripsy forms from printing labels where they where not
needed.
Now crispyforms is out and bootstrap is in. In bootstrap3 we can turn off
unwanted labels easily. It uses the labels however to place placeholders. By
allowing the original labels, thus placeholders are also restored.