Users can scroll after slimscroll state change once more
If a user would click on a link inside of a slimscroll container the
entire page would be blocked from scrolling, this is because slimscroll
set itself on the window element instead of the container element.

We can use an official version of slimscroll again if Pull Request 198
gets accepted.