LILY-2772: Use custom Python base image for HelloLily docker image.
Use custom Python base image for HelloLily docker image, so no more sudden breakage because of a package or base image update (we can control when and how the base image is updated).

Additionally, the custom Dockerfile for the database is removed, since you can just mount the Dockers/db/ folder and Postgres will import the .sql files and execute the .sh files before startup.